untangle spack.config / spack.util.cpus
